+++ lib/secure_window_controller.dart
@@ -1,6 +1,10 @@
+import 'dart:async';
+
 import 'package:flutter/foundation.dart';
 import 'package:secure_window/secure_window.dart';
 import 'package:secure_window/secure_window_state.dart';
+
+enum SecureWindowAuthenticationStatus { SUCCESS, FAILED, NONE }
 
 /// main controller for the library
 ///
@@ -12,6 +16,18 @@ import 'package:secure_window/secure_window_state.dart';
 class SecureWindowController extends ValueNotifier<SecureWindowState> {
   SecureWindowController(SecureWindowState value) : super(value);
 
+  final StreamController<SecureWindowAuthenticationStatus>
+      _authenticationEventsController =
+      StreamController<SecureWindowAuthenticationStatus>.broadcast();
+
+  /// Broadcast stream that you can use to trigger succesffull or unsuccessfull event
+  ///
+  /// will trigger with the result of [SecureApllication.onNeedUnlock]
+  /// ```dart
+  /// secureWindowContrller.authentificationEvents.
+  Stream<SecureWindowAuthenticationStatus> get authenticationEvents =>
+      _authenticationEventsController.stream;
+
   /// Is the application Locked
   /// if locked gates will hide their children content
   /// will be automatically set to yes when user switch back to app
@@ -20,6 +36,28 @@ class SecureWindowController extends ValueNotifier<SecureWindowState> {
 
   /// Is the application secured
   bool get secured => value.secured;
+
+  /// notify listener of the [SecureWindowController.authenticationEvents] of a failure or success
+  /// to allow them for example to clear sensitive data
+  void sendAuthenticationEvent(SecureWindowAuthenticationStatus status) {
+    _authenticationEventsController.add(status);
+  }
+
+  void authFailed({bool unlock = false}) {
+    _authenticationEventsController
+        .add(SecureWindowAuthenticationStatus.FAILED);
+    if (unlock) {
+      this.unlock();
+    }
+  }
+
+  void authSuccess({bool unlock = false}) {
+    _authenticationEventsController
+        .add(SecureWindowAuthenticationStatus.SUCCESS);
+    if (unlock) {
+      this.unlock();
+    }
+  }
 
   /// content under [SecureGate] will not be visible
   void lock() {
@@ -64,5 +102,11 @@ class SecureWindowController extends ValueNotifier<SecureWindowState> {
       value = value.copyWith(secured: false);
       notifyListeners();
     }
+  }
+
+  @override
+  void dispose() {
+    _authenticationEventsController.close();
+    super.dispose();
   }
 }