+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+* new stream for authentication status [SecureWindowController.authenticationEvents]
+* new callback for [SecureApplication]: [SecureApplication.onAuthenticationFailed] and [SecureApplication.onAuthenticationSucceed]
+
 ## 1.0.1
 
 * Update README